requirements.txt
python-telegram-bot
pillow
moviepy
paypalrestsdk
razorpay
